move naming to separate package
