+++ pkg/controllers/dynakube/extension/service_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/Dynatrace/dynatrace-operator/pkg/api/scheme/fake"
-	"github.com/Dynatrace/dynatrace-operator/pkg/controllers/dynakube/extension/utils"
+	"github.com/Dynatrace/dynatrace-operator/pkg/controllers/dynakube/extension/servicename"
 	"github.com/Dynatrace/dynatrace-operator/pkg/util/kubeobjects/labels"
 	"github.com/Dynatrace/dynatrace-operator/pkg/util/timeprovider"
 	"github.com/Dynatrace/dynatrace-operator/pkg/version"
@@ -38,6 +38,6 @@ func TestReconciler_prepareService(t *testing.T) {
 func TestFQDNNameGeneration(t *testing.T) {
 	t.Run(`Check FQDN name generation`, func(t *testing.T) {
 		dk := createDynakube()
-		assert.Equal(t, "test-name-extensions-controller.test-namespace", utils.BuildFQDName(dk))
+		assert.Equal(t, "test-name-extensions-controller.test-namespace", servicename.BuildFQDN(dk))
 	})
 }
